Hide neo4j warning of to_df() being experimental
Since we should henceforth for the foreseeable future provide
pandas types return values whenever appropriate we should not
warn that's it's experimental in neo4j.
If the neo4j API is change in terms of to_df() we will fix it
internally and the user should not have to worry about things
being experimental in certain packages.

TL;DR: pandas support is not experimental in graphdatascience.